+++ lib/chef/provider/package/yum/yum-dump.py
